+++ Files/Changelog.md
@@ -1,4 +1,4 @@
-//SET DEBUG TO FALSE AGAIN
+//CHANGE VERSION && SET DEBUG TO FALSE AGAIN
 
 4.1.4: 30.04.2020
     -Refractored league Detection